sanitizer/requirements.txt
PyYAML == 6.0
pymongo == 3.12
kink == 0.6.6
boto3 == 1.26.91
boto3-stubs[sns,sqs] == 1.26.91
botocore == 1.29.91
aws_error_utils == 2.7.0
expiringdict == 1.2.2
